ortobella_extracted/nested/requirements.txt
Flask>=2.0,<4.0
requests>=2.0
python-dotenv>=1.0
gunicorn>=20.0
